Bring .NET 9 to the main branch

+++ src/Dassie/CodeGeneration/FileCompiler.cs
@@ -1,18 +1,16 @@
 ﻿using Antlr4.Runtime;
 using Antlr4.Runtime.Tree;
-using Dassie.Configuration;
 using Dassie.CLI;
+using Dassie.Configuration;
 using Dassie.Errors;
 using Dassie.Lowering;
 using Dassie.Parser;
 using Dassie.Text.FragmentStore;
 using Dassie.Validation;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.Remoting.Messaging;
 
 namespace Dassie.CodeGeneration;
 
@@ -76,7 +74,12 @@ public static class FileCompiler
 
         if (!config.KeepIntermediateFiles)
         {
-            File.Delete(intermediatePath);
+            if (File.Exists(intermediatePath))
+                File.Delete(intermediatePath);
+
+            foreach (string file in Directory.GetFiles(".temp"))
+                File.Delete(file);
+
             Directory.Delete(".temp");
         }
